projects/ai_interviewer project-final/ai_interview_production_saas/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pdfplumber
spacy
scikit-learn
pandas
python-dotenv
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
